refactor: move file to utils
